+++ packages/core/src/helpers/articles.ts
@@ -5,8 +5,8 @@
 import { ApolloError } from "..";
 import { PantheonClient } from "../core/pantheon-client";
 import {
+  generateArticleQuery,
   generateListArticlesGQL,
-  GET_ARTICLE_QUERY,
   GET_RECOMMENDED_ARTICLES_QUERY,
   LIST_PAGINATED_ARTICLES_QUERY,
   LIST_PAGINATED_ARTICLES_QUERY_W_CONTENT,
@@ -203,6 +203,9 @@ export async function getArticle(
   client: PantheonClient,
   id: number | string,
   args?: ArticleQueryArgs,
+  related?: {
+    site?: boolean;
+  },
 ) {
   const {
     contentType: requestedContentType,
@@ -212,7 +215,7 @@ export async function getArticle(
   const contentType = buildContentType(requestedContentType);
 
   const article = await client.apolloClient.query({
-    query: GET_ARTICLE_QUERY,
+    query: generateArticleQuery({ withSite: related?.site }),
     variables: {
       id: id.toString(),
       contentType,
@@ -230,6 +233,9 @@ export async function getArticleBySlug(
   client: PantheonClient,
   slug: string,
   args?: ArticleQueryArgs,
+  related?: {
+    site?: boolean;
+  },
 ) {
   const {
     contentType: requestedContentType,
@@ -239,7 +245,7 @@ export async function getArticleBySlug(
   const contentType = buildContentType(requestedContentType);
 
   const article = await client.apolloClient.query({
-    query: GET_ARTICLE_QUERY,
+    query: generateArticleQuery({ withSite: related?.site }),
     variables: {
       slug,
       contentType,
@@ -257,11 +263,19 @@ export async function getArticleBySlugOrId(
   client: PantheonClient,
   slugOrId: number | string,
   args?: ArticleQueryArgs,
+  related?: {
+    site?: boolean;
+  },
 ) {
   // First attempt to retrieve by slug, and fallback to by id if the matching slug
   // couldn't be found.
   try {
-    const article = await getArticleBySlug(client, slugOrId.toString(), args);
+    const article = await getArticleBySlug(
+      client,
+      slugOrId.toString(),
+      args,
+      related,
+    );
 
     if (article) {
       return article;
@@ -273,7 +287,7 @@ export async function getArticleBySlugOrId(
   }
 
   try {
-    const article = await getArticle(client, slugOrId, args);
+    const article = await getArticle(client, slugOrId, args, related);
     return article;
   } catch (e) {
     if (