+++ packages/core/src/helpers/convenience.ts
@@ -113,15 +113,18 @@ async function getAllArticlesWithSummary(
 
 async function getArticleBySlugOrId(
   id: number | string,
-  publishingLevel: "PRODUCTION" | "REALTIME" = "PRODUCTION",
+  args?: Parameters<typeof _getArticleBySlugOrId>[2],
+  related?: Parameters<typeof _getArticleBySlugOrId>[3],
 ) {
   const post = await _getArticleBySlugOrId(
     buildPantheonClient({ isClientSide: false }),
     id,
     {
-      publishingLevel,
+      publishingLevel: "PRODUCTION",
       contentType: "TREE_PANTHEON_V2",
+      ...args,
     },
+    related,
   );
 
   return post;