+++ starters/nextjs-starter-approuter-ts/app/articles/[...uri]/clientside-articleview.tsx
@@ -5,6 +5,7 @@ import {
   PantheonProvider,
   PCCConvenienceFunctions,
   updateConfig,
+  type PublishingLevel,
 } from "@pantheon-systems/pcc-react-sdk";
 import ArticleView from "../../../components/article-view";
 
@@ -17,9 +18,11 @@ updateConfig({
 export const ClientsideArticleView = ({
   article,
   grant,
+  publishingLevel,
 }: {
   article: Article;
   grant?: string | undefined;
+  publishingLevel: keyof typeof PublishingLevel;
 }) => {
   return (
     <PantheonProvider
@@ -28,7 +31,7 @@ export const ClientsideArticleView = ({
         pccGrant: grant,
       })}
     >
-      <ArticleView article={article} />
+      <ArticleView article={article} publishingLevel={publishingLevel} />
     </PantheonProvider>
   );
 };