+++ starters/nextjs-starter-approuter-ts/components/article-view.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useArticle } from "@pantheon-systems/pcc-react-sdk";
-import type { Article } from "@pantheon-systems/pcc-react-sdk";
+import type { Article, PublishingLevel } from "@pantheon-systems/pcc-react-sdk";
 import { ArticleRenderer } from "@pantheon-systems/pcc-react-sdk/components";
 import Image from "next/image";
 import Link from "next/link";
@@ -52,6 +52,7 @@ const componentOverrideMap = {
 type ArticleViewProps = {
   article: Article;
   onlyContent?: boolean;
+  publishingLevel: keyof typeof PublishingLevel;
 };
 
 const ArticleHeader = ({
@@ -105,7 +106,10 @@ const ArticleHeader = ({
   );
 };
 
-export function StaticArticleView({ article, onlyContent }: ArticleViewProps) {
+export function StaticArticleView({
+  article,
+  onlyContent,
+}: Pick<ArticleViewProps, "article" | "onlyContent">) {
   const seoMetadata = getSeoMetadata(article);
 
   return (
@@ -122,7 +126,7 @@ export function StaticArticleView({ article, onlyContent }: ArticleViewProps) {
         }}
       />
 
-      <div className="border-base-300 mt-16 flex w-full gap-x-3 border-t-[1px] pt-7 lg:mt-32">
+      <div className="border-base-300 mt-16 flex w-full flex-wrap gap-x-3 gap-y-3 border-t-[1px] pt-9 lg:mt-32">
         {seoMetadata.keywords != null
           ? (Array.isArray(seoMetadata.keywords)
               ? seoMetadata.keywords
@@ -130,7 +134,7 @@ export function StaticArticleView({ article, onlyContent }: ArticleViewProps) {
             ).map((x, i) => (
               <div
                 key={i}
-                className="text-bold text-neutral-content rounded-full border-[1px] border-[#d4d4d4] bg-[#F5F5F5] px-3 py-1 text-sm !no-underline"
+                className="text-bold text-neutral-content inline-block rounded-full border border-[#D4D4D4] bg-[#F5F5F5] px-3 py-1 text-sm !no-underline"
               >
                 {x}
               </div>
@@ -144,15 +148,16 @@ export function StaticArticleView({ article, onlyContent }: ArticleViewProps) {
 export default function ArticleView({
   article,
   onlyContent,
+  publishingLevel,
 }: ArticleViewProps) {
   const { data } = useArticle(
     article.id,
     {
-      publishingLevel: article.publishingLevel,
+      publishingLevel,
       contentType: "TREE_PANTHEON_V2",
     },
     {
-      skip: article.publishingLevel !== "REALTIME",
+      skip: publishingLevel !== "REALTIME",
     },
   );