+++ starters/nextjs-starter/pages/articles/[...uri].jsx
@@ -2,15 +2,15 @@ import {
   PantheonProvider,
   PCCConvenienceFunctions,
 } from "@pantheon-systems/pcc-react-sdk";
+import { getArticlePathComponentsFromContentStructure } from "@pantheon-systems/pcc-react-sdk/server";
 import { NextSeo } from "next-seo";
 import queryString from "query-string";
 import ArticleView from "../../components/article-view";
 import Layout from "../../components/layout";
 import { getSeoMetadata } from "../../lib/utils";
 import { pantheonAPIOptions } from "../api/pantheoncloud/[...command]";
-import { getArticlePathComponentsFromContentStructure } from "@pantheon-systems/pcc-react-sdk/server";
 
-export default function ArticlePage({ article, grant }) {
+export default function ArticlePage({ article, grant, publishingLevel }) {
   const seoMetadata = getSeoMetadata(article);
 
   return (
@@ -28,7 +28,7 @@ export default function ArticlePage({ article, grant }) {
         />
 
         <div className="prose mx-4 mt-16 text-black sm:mx-6 md:mx-auto">
-          <ArticleView article={article} />
+          <ArticleView article={article} publishingLevel={publishingLevel} />
         </div>
       </Layout>
     </PantheonProvider>
@@ -44,10 +44,7 @@ export async function getServerSideProps({
 
   // Fetch the article and site in parallel
   const [article, site] = await Promise.all([
-    PCCConvenienceFunctions.getArticleBySlugOrId(
-      slugOrId,
-      publishingLevel ? publishingLevel.toString().toUpperCase() : "PRODUCTION",
-    ),
+    PCCConvenienceFunctions.getArticleBySlugOrId(slugOrId),
     PCCConvenienceFunctions.getSite(),
   ]);
 
@@ -67,16 +64,16 @@ export async function getServerSideProps({
   if (
     // Check if the article has a slug
     ((article.slug?.trim().length &&
-    // Check if the slug is not the same as the slugOrId
-    article.slug.toLowerCase() !== slugOrId?.trim().toLowerCase()) ||
-    // Check if the article path is not the same as the uri
-    articlePath.length !== (uri.length - 1) ||
-    // Check if the article (with all the components together) path is not the same as the uri
-    articlePath.join("/") !== uri.slice(0, -1).join("/")) &&
+      // Check if the slug is not the same as the slugOrId
+      article.slug.toLowerCase() !== slugOrId?.trim().toLowerCase()) ||
+      // Check if the article path is not the same as the uri
+      articlePath.length !== uri.length - 1 ||
+      // Check if the article (with all the components together) path is not the same as the uri
+      articlePath.join("/") !== uri.slice(0, -1).join("/")) &&
     // Check if resolvePath in pantheon API options is not null
     pantheonAPIOptions.resolvePath != null
   ) {
-    // If the article was accessed by the id rather than the slug 
+    // If the article was accessed by the id rather than the slug
     // or the article path is not the same as the uri - then redirect to the canonical
     // link (mostly for SEO purposes than anything else).
     return {
@@ -93,7 +90,8 @@ export async function getServerSideProps({
   return {
     props: {
       article,
-      grant,
+      grant: grant || null,
+      publishingLevel: publishingLevel || null,
       recommendedArticles: await PCCConvenienceFunctions.getRecommendedArticles(
         article.id,
       ),