+++ packages/react-sample-library/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pantheon-systems/pcc-vue-sdk
 
+## 3.13.1
+
+### Minor Changes
+
+- Bump to 3.13.1
+
+### Patch Changes
+
+- Updated dependencies [4317fcd]
+- Updated dependencies
+- Updated dependencies [a235787]
+  - @pantheon-systems/pcc-react-sdk@3.13.1
+  - @pantheon-systems/pcc-sdk-core@3.13.1
+
 ## 3.12.1
 
 ### Patch Changes