+++ starters/nextjs-starter-approuter-ts/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pantheon-systems/next-pcc-starter
 
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies [4317fcd]
+- Updated dependencies
+- Updated dependencies [a235787]
+  - @pantheon-systems/pcc-react-sdk@3.13.1
+
 ## 1.2.3
 
 ### Patch Changes